+++ hs-opentelemetry/api/src/OpenTelemetry/Exporter.hs
@@ -1,16 +1,24 @@
 {-# LANGUAGE PatternSynonyms #-}
 
+{- |
+Module      : OpenTelemetry.Exporter
+Description : Re-exports of span exporter types.
+Stability   : experimental
+
+This module is deprecated; prefer 'OpenTelemetry.Exporter.Span'.
+-}
 module OpenTelemetry.Exporter
   {-# DEPRECATED "use OpenTelemetry.Exporter.Span instead" #-} (
   Exporter,
-  SpanExporter (Exporter, exporterExport, exporterShutdown),
+  mkExporter,
+  SpanExporter (..),
   ExportResult (..),
 ) where
 
 import Data.HashMap.Strict (HashMap)
 import Data.Vector (Vector)
 import OpenTelemetry.Exporter.Span
-import OpenTelemetry.Internal.Common.Types (InstrumentationLibrary)
+import OpenTelemetry.Internal.Common.Types (FlushResult (..), InstrumentationLibrary, ShutdownResult (..))
 import OpenTelemetry.Internal.Trace.Types (ImmutableSpan)
 
 
@@ -20,9 +28,11 @@ import OpenTelemetry.Internal.Trace.Types (ImmutableSpan)
 type Exporter a = SpanExporter
 
 
-pattern Exporter :: (HashMap InstrumentationLibrary (Vector ImmutableSpan) -> IO ExportResult) -> IO () -> Exporter ImmutableSpan
-pattern Exporter {exporterExport, exporterShutdown} =
+{-# DEPRECATED mkExporter "use SpanExporter constructor directly" #-}
+mkExporter :: (HashMap InstrumentationLibrary (Vector ImmutableSpan) -> IO ExportResult) -> IO () -> Exporter ImmutableSpan
+mkExporter export shutdown =
   SpanExporter
-    { spanExporterExport = exporterExport
-    , spanExporterShutdown = exporterShutdown
+    { spanExporterExport = export
+    , spanExporterShutdown = shutdown >> pure ShutdownSuccess
+    , spanExporterForceFlush = pure FlushSuccess
     }