+++ hs-opentelemetry/exporters/otlp/src/OpenTelemetry/Exporter/OTLP.hs
@@ -1,7 +1,10 @@
-module OpenTelemetry.Exporter.OTLP
-  {-# DEPRECATED "use OpenTelemetry.Exporter.OTLP.Span instead" #-} (
+module OpenTelemetry.Exporter.OTLP (
   module OpenTelemetry.Exporter.OTLP.Span,
+  module OpenTelemetry.Exporter.OTLP.Metric,
+  module OpenTelemetry.Exporter.OTLP.LogRecord,
 ) where
 
+import OpenTelemetry.Exporter.OTLP.LogRecord
+import OpenTelemetry.Exporter.OTLP.Metric
 import OpenTelemetry.Exporter.OTLP.Span